+++ gulpfile.js
@@ -4,71 +4,73 @@
  */
 
 import gulp from 'gulp';
-const { src, dest, parallel, series, watch } = gulp
 
 // Styles
-import { scss, cssCompress } from './.gulp/styles.js'
+import { scss, cssCompress } from './.gulp/styles.js';
 
 // JavaScript
-import { roll, scripts, compressJS } from './.gulp/javascript.js'
+import { roll, scripts, compressJS } from './.gulp/javascript.js';
 
 // HTML
-import { htmlGenerate, htmlRefresh, htmlCompress, validateHtml, testHtml } from './.gulp/html.js'
+import { htmlGenerate, htmlRefresh, htmlCompress, validateHtml, testHtml, htmlPagesPreview } from './.gulp/html.js';
 
 // Images
-import { imagesCompress, webpCompress, genSvgSprite, imgCopy } from './.gulp/images.js'
+import { imagesCompress, webpCompress, genSvgSprite, imgCopy } from './.gulp/images.js';
 
 // Server
-import { openServer, openBrowser, bumper, clean, cleanDist, lt } from './.gulp/server.js'
+import { openServer, openBrowser, bumper, clean, cleanDist, lt } from './.gulp/server.js';
 
 // Misc
-import { copyFiles, copyBuildFiles, copyVideo, copyFonts, copyIcons } from './.gulp/misc.js'
+import { copyFiles, copyBuildFiles, copyVideo, copyFonts, copyIcons } from './.gulp/misc.js';
 
+// Tests
+import { mobileTestRes, htmlSpeedRes, cssTestRes } from './.gulp/tests.js';
+
+/**
+ * @todo Tasks
+ * @todo Statistics
+ */
 
 /**
  * System
  */
-import concat from 'gulp-concat'
-
+import concat from 'gulp-concat';
+const { src, dest, parallel, series, watch } = gulp;
 
 /**
  * Settings
  * ================================================================================
  */
 
-
 /**
  * JS Libraries appending
  */
 const jsVendorList = {
-  src: ['./dist/javascript/app.js'],
-}
+  src: ['./dist/javascript/app.js']
+};
 
 const jsVendorLibs = () =>
   src(jsVendorList.src)
     .pipe(concat('app.js'))
-    .pipe(dest('./dist/javascript/'))
-
+    .pipe(dest('./dist/javascript/'));
 
 /**
  * Tasks
  * ================================================================================
  */
 
-
 /**
  * Watcher
  */
 const watcher = () => {
-  watch('./src/scss/**/*.scss', scss)
-  watch('./src/**/*.html', series(htmlGenerate, cleanDist, htmlRefresh, testHtml))
-  watch('./src/javascript/**/*.js', series(series(roll, jsVendorLibs), scripts))
-  watch('./src/images/**/*', imgCopy)
-  watch('./src/favicons/**/*', copyIcons)
-  watch('./src/fonts/**/*', copyFonts)
-  watch('./src/video/**/*', copyVideo)
-}
-
+  watch('./src/scss/**/*.scss', scss);
+  watch('./src/**/*.html', series(htmlGenerate, cleanDist, htmlRefresh, testHtml));
+  watch('./src/javascript/**/*.js', series(series(roll, jsVendorLibs), scripts));
+  watch('./src/images/**/*', imgCopy);
+  watch('./src/favicons/**/*', copyIcons);
+  watch('./src/fonts/**/*', copyFonts);
+  watch('./src/video/**/*', copyVideo);
+};
 
 /**
  * Default Tasks
@@ -82,18 +84,17 @@ export default series(
     series(series(roll, jsVendorLibs)),
     series(scss),
     series(imgCopy),
-    series(htmlGenerate, cleanDist, openBrowser, validateHtml),
+    series(htmlGenerate, cleanDist, openBrowser, htmlPagesPreview, validateHtml),
     openServer,
     lt,
     watcher
   )
-)
+);
 
 /**
  * Test Tasks
  */
-
-const test = parallel(validateHtml)
+const test = parallel(mobileTestRes, htmlSpeedRes, cssTestRes);
 
 /**
  * Build Tasks
@@ -108,19 +109,16 @@ const build = series(
     series(htmlGenerate, cleanDist, htmlCompress)
   ),
   bumper
-)
+);
 
 /**
  * Generate Compressed Images
  */
-const images = series(imagesCompress, webpCompress)
+const images = series(imagesCompress, webpCompress);
 
 /**
  * Generate SVG Sprite
  */
-const svgSprite = series(genSvgSprite)
+const sprite = series(genSvgSprite);
 
-// Generate HTML
-const html = series(htmlGenerate, cleanDist)
-
-export { test, build, images, svgSprite, html }
+export { test, build, images, sprite };