+++ src/javascript/app.js
@@ -2,15 +2,14 @@
  * Load methods, helpers, polyfills etc.
  */
 
-import domReady from 'js-components-npm/src/utils/domready'
-import Parallax from 'js-components-npm/src/utils/parallax'
-import NoJs from 'js-components-npm/src/utils/no-js'
-
-import ModuleClass from 'src/module-class'
-import Component from 'src/components/component'
+import domReady from 'js-components-npm/src/utils/domready';
+// import Parallax from 'js-components-npm/src/utils/parallax';
+import NoJs from 'js-components-npm/src/utils/no-js';
+import ModuleClass from 'src/module-class';
+import Component from 'src/components/component';
 
 // Check JS is enabled
-NoJs()
+NoJs();
 
 /**
  * The DOMContentLoaded event fires when the initial HTML document has been completely loaded and parsed,
@@ -19,9 +18,9 @@ NoJs()
  * https://developer.mozilla.org/en-US/docs/Web/API/Window/DOMContentLoaded_event
  */
 domReady(() => {
-  Component()
+  Component();
 
   // Test Class
-  const obj = new ModuleClass()
-  obj.method()
-})
+  const obj = new ModuleClass();
+  obj.method();
+});